Caesar Cipher and some stuff

+++ Experimental/FibonacciSequence/FibonacciSequence/FibonacciSequence.h
@@ -4,7 +4,7 @@
 
 void FibonacciSequence(int n) {
 
-	unsigned int t1 = 0, t2 = 1, temp;
+	unsigned long long t1 = 0, t2 = 1, temp;
 
 	for (int i = 0; i < n; i++) {
 		temp = t1 + t2;
